let ClassDefCreator dump @@schema_element in multiple-and-indented lines. contains partial fixes for #168.

+++ test/wsdl/ref/test_ref.rb
@@ -23,7 +23,7 @@ class TestRef < Test::Unit::TestCase
       gen.run
     end
     compare("expectedProduct.rb", "product.rb")
-    File.unlink(pathname('product.rb'))
+    File.unlink(pathname('product.rb')) unless $DEBUG
   end
 
   def compare(expected, actual)
